+++ addon/mixins/select-box/registration/has-selected-options.js
@@ -1,31 +1,19 @@
 import Mixin from '@ember/object/mixin';
 import { A as emberA } from '@ember/array';
-import { scheduleOnce } from '@ember/runloop';
 
 export default Mixin.create({
   init() {
     this._super(...arguments);
-    this.set('_selectedOptions', emberA());
     this.set('selectedOptions', emberA());
-  },
-
-  _scheduleUpdateSelectedOptions() {
-    scheduleOnce('afterRender', this, '_updateSelectedOptions');
-  },
-
-  _updateSelectedOptions() {
-    this.set('selectedOptions', emberA(this.get('_selectedOptions').toArray()));
   },
 
   actions: {
     _registerSelectedOption(selectedOption) {
-      this.get('_selectedOptions').addObject(selectedOption);
-      this._scheduleUpdateSelectedOptions();
+      this.get('selectedOptions').addObject(selectedOption);
     },
 
     _deregisterSelectedOption(selectedOption) {
-      this.get('_selectedOptions').removeObject(selectedOption);
-      this._scheduleUpdateSelectedOptions();
+      this.get('selectedOptions').removeObject(selectedOption);
     }
   }
 });